Manually remove some annotations

+++ exercises/difference-of-squares/difference-of-squares.jl
@@ -1,14 +1,14 @@
 "Square the sum of the first `n` positive integers"
-function square_of_sum(n::Int)
+function square_of_sum(n)
 
 end
 
 "Sum the squares of the first `n` positive integers"
-function sum_of_squares(n::Int)
+function sum_of_squares(n)
 
 end
 
 "Subtract the sum of squares from square of the sum of the first `n` positive ints"
-function difference(n::Int)
+function difference(n)
 
 end
